pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "devagent-cli"
version = "0.1.0"
description = "Local-first agentic AI developer assistant CLI"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "DevAgent CLI" }
]
dependencies = [
  "typer>=0.12.0",
  "rich>=13.7.0",
  "google-genai>=1.0.0",
  "gitpython>=3.1.43",
  "watchdog>=4.0.0",
  "chromadb>=0.5.0",
  "python-dotenv>=1.0.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0"
]

[project.scripts]
devagent = "devagent.cli.main:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.hatch.build.targets.wheel]
packages = ["devagent"]
